+++ modules/core/dodecahedron-quaternions.ts
